requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Login==0.2.11
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
enum34==1.0.4
hashids==1.0.1
itsdangerous==0.24
passlib==1.6.2
pathlib==1.0.1
pyScss==1.3.4
pymongo==2.7.2
pyparsing==2.0.3
requests==2.5.0
six==1.8.0
webassets==0.10.1
wsgiref==0.1.2
